requirements.txt
boto3
pyramid==2.0
mysql-connector-python